# openapi/urbanbloom-api-v1.yaml
openapi: 3.0.3
info:
  title: UrbanBloom API
  version: 1.0.0
  description: |
    **UrbanBloom** is a mobile and web application for documenting and gamifying environmental
    actions in cities. This API follows a Domain-Driven Design (DDD) approach with 9 bounded contexts.

    ## Architecture
    - RESTful API with OpenAPI 3.0 specification
    - JWT-based authentication (Keycloak)
    - Event-driven communication between bounded contexts
    - Offline-first support via sync endpoints

servers:
  - url: http://localhost:8080/api/v1
    description: Local development server

tags:
  - name: Authentication
    description: User registration and authentication
  - name: Users
    description: User profile management

security:
  - BearerAuth: []

paths:
  # ===== 1. User / Identity Context =====

  /registration:
    post:
      tags: [Authentication]
      summary: Register a new citizen
      description: Creates a new citizen account (Keycloak + PostgreSQL)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationRequest'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/mobile/login:
    post:
      tags: [Authentication]
      summary: Login for mobile citizens
      description: Authenticate citizen via Keycloak Direct Grant
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials or account not verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/admin/login:
    post:
      tags: [Authentication]
      summary: Login for administrators
      description: Authenticate admin via Keycloak Direct Grant
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Password change required (first login)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdatePasswordRequiredResponse'
        '403':
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/logout:
    post:
      tags: [Authentication]
      summary: User logout
      description: Revoke refresh token in Keycloak
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /auth/password/reset-request:
    post:
      tags: [Authentication]
      summary: Request password reset
      description: Trigger Keycloak password reset email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Always returns OK (neutral feedback)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /users/me:
    get:
      tags: [Users]
      summary: Get current user profile
      description: Fetch enriched user profile from PostgreSQL
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'

# (rest of the OpenAPI paths for other contexts could be kept or added back later, 
# but for MS-01 focus on Identity)

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string

    RegistrationRequest:
      type: object
      required: [email, password, firstName, lastName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        firstName:
          type: string
        lastName:
          type: string

    RegistrationResponse:
      type: object
      required: [userId, externalId, message]
      properties:
        userId:
          type: string
          format: uuid
        externalId:
          type: string
        message:
          type: string

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password

    LoginResponse:
      type: object
      required: [accessToken, refreshToken, expiresIn]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
        tokenType:
          type: string
          default: Bearer

    UpdatePasswordRequiredResponse:
      type: object
      required: [error, message, keycloakLoginUrl]
      properties:
        error:
          type: string
          example: UPDATE_PASSWORD_REQUIRED
        message:
          type: string
        keycloakLoginUrl:
          type: string

    UserDetail:
      type: object
      required: [id, externalId, email, firstName, lastName, status, roles, registeredAt]
      properties:
        id:
          type: string
          format: uuid
        externalId:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        status:
          type: string
          enum: [ENABLED, DISABLED]
        roles:
          type: array
          items:
            type: string
        registeredAt:
          type: string
          format: date-time